refactor: 동일한 Url의 서버에서 데이터를 받아오는 경우 Service 하나만 생성
